+++ app/controllers/sales_controller.rb
@@ -12,7 +12,7 @@ class SalesController < ApplicationController
    @sale.lead = @current_lead
    @sale.date = Time.now
    if @sale.save
-     Recall.where(lead: @current_lead).first.destroy
+     Recall.where(lead: @current_lead).first.destroy if Recall.where(lead: @current_lead).count > 0
      flash[:notice] = "#{current_user.nickname} sold #{@current_lead.name1} $#{@sale.amount}."
      redirect_to campaign_lead_path(@current_lead.campaign, @next_lead)
    else
